refactor(core-api): extract common domain services scaffold
