+++ apps/core-api/src/auth/auth.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthGuard } from './auth.guard';
-import { PrismaService } from '../prisma/prisma.service';
 import { AuthController } from './auth.controller';
+import { CommonServicesModule } from '../common/common-services.module';
 
 @Module({
+  imports: [CommonServicesModule],
   controllers: [AuthController],
-  providers: [AuthService, AuthGuard, PrismaService],
-  exports: [AuthService, AuthGuard, PrismaService],
+  providers: [AuthService, AuthGuard],
+  exports: [AuthService, AuthGuard],
 })
 export class AuthModule {}
