+++ apps/core-api/src/capacity/capacity.module.ts
@@ -1,14 +1,13 @@
 import { Module } from '@nestjs/common';
 import { AuthModule } from '../auth/auth.module';
-import { DomainService } from '../common/domain.service';
-import { PrismaService } from '../prisma/prisma.service';
+import { CommonServicesModule } from '../common/common-services.module';
 import { CapacityController } from './capacity.controller';
 import { CapacityService } from './capacity.service';
 
 @Module({
-  imports: [AuthModule],
+  imports: [AuthModule, CommonServicesModule],
   controllers: [CapacityController],
-  providers: [CapacityService, PrismaService, DomainService],
+  providers: [CapacityService],
   exports: [CapacityService],
 })
 export class CapacityModule {}
