+++ apps/core-api/src/common/domain.service.ts
@@ -1,13 +1,9 @@
-import { Injectable, ForbiddenException, NotFoundException, ConflictException, Inject, BadRequestException } from '@nestjs/common';
-import { PrismaService } from '../prisma/prisma.service';
+import { Injectable, BadRequestException, ConflictException, Inject } from '@nestjs/common';
 import {
-  GuestAccessStatus,
-  GuestAccessScopeType,
   Prisma,
   ProjectRole,
   TaskStatus,
   TaskType,
-  UserStatus,
   WorkspaceRole,
 } from '@prisma/client';
 import {
@@ -21,186 +17,49 @@ import {
   type TaskStatus as DomainTaskStatus,
   type TaskType as DomainTaskType,
 } from '@atlaspm/domain';
-import { templateDefinition } from '../rules/rule-definition';
 import type { AuthUser } from './types';
+import { AuditOutboxService } from './audit-outbox.service';
+import { AuthorizationService } from './authorization.service';
+import { IdentityService } from './identity.service';
+import { WorkspaceDefaultsService } from './workspace-defaults.service';
 
 @Injectable()
 export class DomainService {
   private static readonly defaultRuleTemplateKeys = ['progress_to_done', 'progress_to_in_progress'] as const;
-  private static readonly defaultWorkspaceRetryDelaysMs = [10, 25, 50, 100, 200, 400];
-  private static readonly workspaceRoleRank: Record<WorkspaceRole, number> = { WS_MEMBER: 1, WS_ADMIN: 2 };
-  private static readonly projectRoleRank: Record<ProjectRole, number> = { VIEWER: 1, MEMBER: 2, ADMIN: 3 };
 
-  constructor(@Inject(PrismaService) private readonly prisma: PrismaService) {}
+  constructor(
+    @Inject(AuditOutboxService) private readonly auditOutbox: AuditOutboxService,
+    @Inject(AuthorizationService) private readonly authorization: AuthorizationService,
+    @Inject(IdentityService) private readonly identity: IdentityService,
+    @Inject(WorkspaceDefaultsService) private readonly defaults: WorkspaceDefaultsService,
+  ) {}
 
   isDefaultRuleTemplateKey(templateKey: string): boolean {
     return DomainService.defaultRuleTemplateKeys.includes(templateKey as typeof DomainService.defaultRuleTemplateKeys[number]);
   }
 
   async ensureUser(sub: string, email?: string, name?: string) {
-    const existing = await this.prisma.user.findUnique({ where: { id: sub } });
-    if (!existing) {
-      return this.prisma.user.create({
-        data: { id: sub, email, displayName: name, status: UserStatus.ACTIVE },
-      });
-    }
-    return this.prisma.user.update({
-      where: { id: sub },
-      data: {
-        email,
-        displayName: existing.displayName ?? name,
-      },
-    });
+    return this.identity.ensureUser(sub, email, name);
   }
 
   async syncAuthenticatedUser(user: AuthUser) {
-    const existing = await this.prisma.user.findUnique({ where: { id: user.sub } });
-    const now = new Date();
-    if (!existing) {
-      return this.prisma.user.create({
-        data: {
-          id: user.sub,
-          email: user.email,
-          displayName: user.name,
-          status: UserStatus.ACTIVE,
-          lastSeenAt: now,
-        },
-      });
-    }
-    if (existing.status === UserStatus.SUSPENDED) {
-      throw new ForbiddenException('User is suspended');
-    }
-    return this.prisma.user.update({
-      where: { id: user.sub },
-      data: {
-        email: user.email,
-        displayName: existing.displayName ?? user.name,
-        lastSeenAt: now,
-      },
-    });
+    return this.identity.syncAuthenticatedUser(user);
   }
 
   async ensureDefaultWorkspaceForUser(sub: string) {
-    for (let attempt = 0; attempt < DomainService.defaultWorkspaceRetryDelaysMs.length; attempt += 1) {
-      try {
-        return await this.prisma.$transaction(async (tx) => {
-          let membership = await tx.workspaceMembership.findFirst({
-            where: { userId: sub },
-            include: { workspace: true },
-          });
-
-          if (!membership) {
-            const ws = await tx.workspace.create({ data: { name: 'Default Workspace' } });
-            membership = await tx.workspaceMembership.create({
-              data: { workspaceId: ws.id, userId: sub, role: WorkspaceRole.WS_ADMIN },
-              include: { workspace: true },
-            });
-          }
-
-          return membership.workspace;
-        }, {
-          maxWait: 5000,
-          timeout: 10000,
-          isolationLevel: 'Serializable',
-        });
-      } catch (error) {
-        const maybePrismaError = error as { code?: string; message?: string };
-        const isRetryableTxnConflict =
-          maybePrismaError.code === 'P2034' ||
-          maybePrismaError.message?.includes('write conflict') ||
-          maybePrismaError.message?.includes('deadlock') ||
-          maybePrismaError.message?.includes('serialization');
-
-        if (!isRetryableTxnConflict || attempt === DomainService.defaultWorkspaceRetryDelaysMs.length - 1) {
-          const membership = await this.prisma.workspaceMembership.findFirst({
-            where: { userId: sub },
-            include: { workspace: true },
-          });
-          if (membership) {
-            return membership.workspace;
-          }
-          throw error;
-        }
-
-        const membership = await this.prisma.workspaceMembership.findFirst({
-          where: { userId: sub },
-          include: { workspace: true },
-        });
-        if (membership) {
-          return membership.workspace;
-        }
-
-        const delayMs = DomainService.defaultWorkspaceRetryDelaysMs[attempt];
-        await new Promise((resolve) => setTimeout(resolve, delayMs));
-      }
-    }
-
-    throw new ConflictException('Failed to ensure default workspace');
+    return this.defaults.ensureDefaultWorkspaceForUser(sub);
   }
 
   async requireWorkspaceRole(workspaceId: string, userId: string, min: WorkspaceRole) {
-    const membership = await this.prisma.workspaceMembership.findUnique({
-      where: { workspaceId_userId: { workspaceId, userId } },
-    });
-    if (!membership) throw new NotFoundException('Workspace membership not found');
-    if (DomainService.workspaceRoleRank[membership.role] < DomainService.workspaceRoleRank[min]) {
-      throw new ForbiddenException('Insufficient workspace role');
-    }
-    return membership;
+    return this.authorization.requireWorkspaceRole(workspaceId, userId, min);
   }
 
   async requireWorkspaceMembership(workspaceId: string, userId: string) {
-    const membership = await this.prisma.workspaceMembership.findUnique({
-      where: { workspaceId_userId: { workspaceId, userId } },
-    });
-    if (!membership) throw new NotFoundException('Workspace membership not found');
-    return membership;
+    return this.authorization.requireWorkspaceMembership(workspaceId, userId);
   }
 
   async requireProjectRole(projectId: string, userId: string, min: ProjectRole) {
-    const membership = await this.prisma.projectMembership.findUnique({
-      where: { projectId_userId: { projectId, userId } },
-    });
-    if (membership) {
-      if (DomainService.projectRoleRank[membership.role] < DomainService.projectRoleRank[min]) {
-        throw new ForbiddenException('Insufficient role');
-      }
-      return membership;
-    }
-
-    const guestGrant = await this.prisma.guestAccessGrant.findFirst({
-      where: {
-        userId,
-        projectId,
-        scopeType: GuestAccessScopeType.PROJECT,
-      },
-      orderBy: { updatedAt: 'desc' },
-    });
-    if (!guestGrant) {
-      throw new NotFoundException('Project membership not found');
-    }
-    if (guestGrant.status === GuestAccessStatus.REVOKED || guestGrant.revokedAt) {
-      throw new ForbiddenException('Guest access has been revoked');
-    }
-    if (
-      guestGrant.status === GuestAccessStatus.EXPIRED ||
-      (guestGrant.expiresAt && guestGrant.expiresAt.getTime() <= Date.now())
-    ) {
-      throw new ForbiddenException('Guest access has expired');
-    }
-    if (!guestGrant.projectRole) {
-      throw new ForbiddenException('Guest project access is not configured');
-    }
-    if (DomainService.projectRoleRank[guestGrant.projectRole] < DomainService.projectRoleRank[min]) {
-      throw new ForbiddenException('Insufficient role');
-    }
-    return {
-      id: guestGrant.id,
-      projectId,
-      userId,
-      role: guestGrant.projectRole,
-      createdAt: guestGrant.createdAt,
-    };
+    return this.authorization.requireProjectRole(projectId, userId, min);
   }
 
   async appendAuditOutbox(args: {
@@ -215,93 +74,11 @@ export class DomainService {
     outboxType: string;
     payload: unknown;
   }) {
-    const correlationId = args.correlationId ?? 'test-correlation-id';
-    const beforeJson =
-      args.beforeJson === undefined
-        ? undefined
-        : args.beforeJson === null
-          ? Prisma.JsonNull
-          : (args.beforeJson as Prisma.InputJsonValue);
-    const afterJson =
-      args.afterJson === undefined
-        ? undefined
-        : args.afterJson === null
-          ? Prisma.JsonNull
-          : (args.afterJson as Prisma.InputJsonValue);
-    const payload =
-      args.payload === null ? Prisma.JsonNull : (args.payload as Prisma.InputJsonValue);
-
-    await args.tx.auditEvent.create({
-      data: {
-        actor: args.actor,
-        entityType: args.entityType,
-        entityId: args.entityId,
-        action: args.action,
-        beforeJson,
-        afterJson,
-        correlationId,
-      },
-    });
-    await args.tx.outboxEvent.create({
-      data: {
-        type: args.outboxType,
-        payload,
-        correlationId,
-      },
-    });
+    return this.auditOutbox.appendAuditOutbox(args);
   }
 
   async ensureProjectDefaults(projectId: string, actor: string, correlationId: string) {
-    const tx = this.prisma;
-    const defaultSection = await tx.section.findFirst({ where: { projectId, isDefault: true } });
-    if (!defaultSection) {
-      const section = await tx.section.create({
-        data: { projectId, name: 'No Section', isDefault: true, position: 1000 },
-      });
-      await this.appendAuditOutbox({
-        tx,
-        actor,
-        entityType: 'Section',
-        entityId: section.id,
-        action: 'section.created.default',
-        afterJson: section,
-        correlationId,
-        outboxType: 'section.created',
-        payload: section,
-      });
-    }
-
-    const templates = [
-      { name: 'Progress to Done', templateKey: 'progress_to_done' },
-      { name: 'Progress to In Progress', templateKey: 'progress_to_in_progress' },
-    ] as const;
-    for (const tpl of templates) {
-      const rule = await tx.rule.upsert({
-        where: { projectId_templateKey: { projectId, templateKey: tpl.templateKey } },
-        create: {
-          projectId,
-          name: tpl.name,
-          templateKey: tpl.templateKey,
-          definition: templateDefinition(tpl.templateKey) as Prisma.InputJsonValue,
-          enabled: true,
-          cooldownSec: 60,
-        },
-        update: {
-          definition: templateDefinition(tpl.templateKey) as Prisma.InputJsonValue,
-        },
-      });
-      await this.appendAuditOutbox({
-        tx,
-        actor,
-        entityType: 'Rule',
-        entityId: rule.id,
-        action: 'rule.ensure_template',
-        afterJson: rule,
-        correlationId,
-        outboxType: 'rule.created',
-        payload: rule,
-      });
-    }
+    return this.defaults.ensureProjectDefaults(projectId, actor, correlationId);
   }
 
   async ensureProjectDefaultsInTx(
@@ -310,63 +87,7 @@ export class DomainService {
     actor: string,
     correlationId: string,
   ) {
-    const defaultSection = await tx.section.findFirst({ where: { projectId, isDefault: true } });
-    if (!defaultSection) {
-      const section = await tx.section.create({
-        data: { projectId, name: 'No Section', isDefault: true, position: 1000 },
-      });
-      await this.appendAuditOutbox({
-        tx,
-        actor,
-        entityType: 'Section',
-        entityId: section.id,
-        action: 'section.created.default',
-        afterJson: section,
-        correlationId,
-        outboxType: 'section.created',
-        payload: section,
-      });
-    }
-
-    const templates = [
-      { name: 'Progress to Done', templateKey: 'progress_to_done' },
-      { name: 'Progress to In Progress', templateKey: 'progress_to_in_progress' },
-    ] as const;
-    for (const tpl of templates) {
-      const existing = await tx.rule.findUnique({
-        where: { projectId_templateKey: { projectId, templateKey: tpl.templateKey } },
-      });
-      if (existing) {
-        if (!existing.definition) {
-          await tx.rule.update({
-            where: { id: existing.id },
-            data: { definition: templateDefinition(tpl.templateKey) as Prisma.InputJsonValue },
-          });
-        }
-        continue;
-      }
-      const rule = await tx.rule.create({
-        data: {
-          projectId,
-          name: tpl.name,
-          templateKey: tpl.templateKey,
-          definition: templateDefinition(tpl.templateKey) as Prisma.InputJsonValue,
-          enabled: true,
-          cooldownSec: 60,
-        },
-      });
-      await this.appendAuditOutbox({
-        tx,
-        actor,
-        entityType: 'Rule',
-        entityId: rule.id,
-        action: 'rule.ensure_template',
-        afterJson: rule,
-        correlationId,
-        outboxType: 'rule.created',
-        payload: rule,
-      });
-    }
+    return this.defaults.ensureProjectDefaultsInTx(tx, projectId, actor, correlationId);
   }
 
   ensureProgressRange(progressPercent?: number) {
