+++ apps/core-api/src/goals/goals.module.ts
@@ -1,14 +1,13 @@
 import { Module } from '@nestjs/common';
 import { AuthModule } from '../auth/auth.module';
-import { DomainService } from '../common/domain.service';
-import { PrismaService } from '../prisma/prisma.service';
+import { CommonServicesModule } from '../common/common-services.module';
 import { GoalsController } from './goals.controller';
 import { GoalsService } from './goals.service';
 
 @Module({
-  imports: [AuthModule],
+  imports: [AuthModule, CommonServicesModule],
   controllers: [GoalsController],
-  providers: [GoalsService, PrismaService, DomainService],
+  providers: [GoalsService],
   exports: [GoalsService],
 })
 export class GoalsModule {}
