+++ apps/core-api/src/workload/workload.module.ts
@@ -1,14 +1,13 @@
 import { Module } from '@nestjs/common';
 import { AuthModule } from '../auth/auth.module';
 import { CapacityModule } from '../capacity/capacity.module';
-import { DomainService } from '../common/domain.service';
-import { PrismaService } from '../prisma/prisma.service';
+import { CommonServicesModule } from '../common/common-services.module';
 import { WorkloadService } from './workload.service';
 import { WorkloadController } from './workload.controller';
 
 @Module({
-  imports: [AuthModule, CapacityModule],
+  imports: [AuthModule, CapacityModule, CommonServicesModule],
   controllers: [WorkloadController],
-  providers: [PrismaService, DomainService, WorkloadService],
+  providers: [WorkloadService],
 })
 export class WorkloadModule {}
